pyproject.toml
[project]
name = "specvod-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "daphne>=4.2.1",
    "django>=5.2.2",
    "django-ckeditor-5>=0.2.18",
    "django-cors-headers>=4.9.0",
    "django-dbbackup>=5.0.0",
    "django-mptt>=0.18.0",
    "django-ninja>=1.5.0",
    "mysqlclient>=2.2.7",
    "pillow>=12.0.0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "sorl-thumbnail>=12.11.0",
]
